# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: 影策 HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /plugin-operations:
    get:
      tags: [agent]
      summary: 检索当前用户已启用且获准的插件操作
      parameters:
        - {name: q, in: query, schema: {type: string, maxLength: 200}}
        - {name: cursor, in: query, schema: {type: integer, minimum: 0}}
        - {name: hostSurface, in: query, schema: {type: string, enum: [agent-home, canvas, editor]}}
        - {name: canvasId, in: query, schema: {type: string}}
        - {name: projectId, in: query, schema: {type: string}}
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "403":
          $ref: "#/components/responses/Error"
  /plugin-operations/{pluginId}/{operationId}:
    get:
      tags: [agent]
      summary: 读取固定发布版本的操作及输入输出 Schema
      parameters:
        - {name: pluginId, in: path, required: true, schema: {type: string}}
        - {name: operationId, in: path, required: true, schema: {type: string}}
        - {name: releaseId, in: query, required: true, schema: {type: string}}
        - {name: hostSurface, in: query, schema: {type: string, enum: [agent-home, canvas, editor]}}
        - {name: canvasId, in: query, schema: {type: string}}
        - {name: projectId, in: query, schema: {type: string}}
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /plugin-invocations:
    post:
      tags: [agent]
      summary: 调用已启用插件操作
      description: 纯读取同步返回 inline；持久写操作必须带 Idempotency-Key，返回 waiting_approval 的 PluginRun，不由 Agent 自批。
      parameters:
        - {name: Idempotency-Key, in: header, schema: {type: string, minLength: 8, maxLength: 160}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operation, releaseId, input]
              additionalProperties: false
              properties:
                operation: {type: string}
                releaseId: {type: string}
                input: {type: object}
                context:
                  type: object
                  additionalProperties: false
                  properties:
                    hostSurface: {type: string, enum: [agent-home, canvas, editor]}
                    canvasId: {type: string}
                    projectId: {type: string}
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /plugin-runs/{id}:
    get:
      tags: [agent]
      summary: 读取当前用户的单步插件运行
      parameters:
        - $ref: "#/components/parameters/AgentRunID"
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "404":
          $ref: "#/components/responses/Error"
  /plugin-runs/{id}/approvals/{approvalId}:
    post:
      tags: [agent]
      summary: 用户批准或拒绝持久插件操作
      parameters:
        - $ref: "#/components/parameters/AgentRunID"
        - {name: approvalId, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision, revision]
              additionalProperties: false
              properties:
                decision: {type: string, enum: [approve, reject]}
                revision: {type: integer, minimum: 1}
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "409":
          $ref: "#/components/responses/Error"
  /plugin-runs/{id}/cancel:
    post:
      tags: [agent]
      summary: 取消当前用户仍在等待确认的插件运行
      parameters:
        - $ref: "#/components/parameters/AgentRunID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revision]
              additionalProperties: false
              properties:
                revision: {type: integer, minimum: 1}
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    AgentRunID:
      name: id
      in: path
      required: true
      schema:
        type: string
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
...
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - bad_gateway
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
  responses:
...
    Envelope:
      description: 业务成功信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"